separate report-generator and app logic
